src/chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "shared",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "langchain>=0.2",
    "langchain-openai>=0.1",
    "langgraph>=0.2",
    "falkordb>=1.0",
    "langchain-core>=0.2",
    "litellm>=1.40",
    "sentence-transformers>=3.0",
    "structlog>=24.1",
    "websockets>=12.0",
    "python-jose[cryptography]>=3.3",
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0.30",
]

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chatbot"]
